Bits and pieces.

+++ src/maths.rs
@@ -28,14 +28,14 @@ const EPSILON: f32 = 0.00001;
 
 /// Converts the given value to radians from degrees.
 #[inline]
-pub fn to_radians(degrees: f64) -> f64 {
-  degrees * (std::f64::consts::PI / 180.0)
+pub fn to_radians(degrees: f32) -> f32 {
+  degrees * (std::f32::consts::PI / 180.0)
 }
 
 /// Converts the given value to degrees to radians.
 #[inline]
-pub fn to_degrees(radians: f64) -> f64 {
-  (radians * 180.0) / std::f64::consts::PI
+pub fn to_degrees(radians: f32) -> f32 {
+  (radians * 180.0) / std::f32::consts::PI
 }
 
 /// Allows approximate equality checks between values.
